requirements.txt
tensorflow
fastapi
uvicorn
pillow
numpy
matplotlib
scikit-learn
python-multipart
pandas

